cmd/audio: Implement a queue system
If the play sub-command is executed while an item is already playing
then the new item will be queued and start playing when the current one
has finished.

The stop sub-command was replaced by the skip sub-command since simply
stopping doesn't make a lot of sense since there's a queue.