cmd/paintball: Fix crash when shuffling title
The shuffling function used runes to shuffle but counted the string's
length using the generic len() function which just counts the bytes.
This means that if a there was a difference between the number of bytes
and the number of runes (for example with a title like 'WALL·E') the bot
would crash.